+++ src/jflow/db/instdata/tests/forms.py
@@ -1,8 +1,11 @@
+from django.conf import settings
 from django.test import TestCase
 from django.contrib.contenttypes.models import ContentType
+from django.contrib.auth.models import User
 from jflow.db.instdata.forms import DataIdForm
 from jflow.db.instdata import models
 
+ADMIN_URL_PREFIX = getattr(settings, 'ADMIN_URL_PREFIX', '/admin/') 
 
 class FormTests(TestCase):
     fixtures = ['initial_data.json',
@@ -11,6 +14,21 @@ class FormTests(TestCase):
     def setUp(self):
         self.ct = models.CollateralType(name = 'collateral')
         self.ct.save()
+        self.su = User.objects.create_superuser('superuser', 'test@example.com', 'superpw')
+        self.client.login(username='superuser', password='superpw')
+    
+    def bond_data(self, code = 'A_BOND_CODE', country = 'GE', curncy = 'EUR'):
+        ct = ContentType.objects.get_for_model(models.Bond)
+        return {'code': code,
+                'country': country,
+                'bond_class': 1,
+                'collateral_type': self.ct.pk,
+                'content_type': ct.pk,
+                'day_count': 'act360',
+                'month_frequency': 12,
+                'multiplier': 0.01,
+                'settlement_delay': 3,
+                'callable': False}
         
     def create_equity(self, code = 'abcd', country = 'US', curncy = 'EUR', instance = None):
         ct = ContentType.objects.get_for_model(models.Equity)
@@ -25,20 +43,12 @@ class FormTests(TestCase):
         return f.save()
     
     def create_bond(self, code = 'bond', country = 'GE', curncy = 'EUR', instance = None):
-        ct = ContentType.objects.get_for_model(models.Bond)
-        f = DataIdForm({'code': code,
-                        'country': country,
-                        'bond_class': 1,
-                        'collateral_type': self.ct.pk,
-                        'content_type': ct.pk,
-                        'day_count': 'act360',
-                        'month_frequency': 12,
-                        'multiplier': 0.01,
-                        'settlement_delay': 3,
-                        'callable': False},
-                        instance = instance)
+        f = DataIdForm(self.bond_data(code,country,curncy), instance = instance)
         self.assertTrue(f.is_valid())
         return f.save()
+    
+    def get(self, code):
+        return models.DataId.objects.get(code = code)
     
     def testNoInstrument1(self):
         f = DataIdForm({'code': 'spxt','country':'US'})
@@ -87,8 +97,18 @@ class FormTests(TestCase):
         self.assertEqual(nid.type,'bond')
         eqs = models.Equity.objects.filter(code = 'VOD_LN')
         self.assertEqual(eqs.count(),0)
-        
         self.assertEqual(id.type,'bond')
         self.assertEqual(id.instrument.bond_class.code,'DBR')
+        
+    def testAdmin(self):
+        if 'django.contrib.admin' in settings.INSTALLED_APPS:
+            data = self.bond_data(code = 'dbr54')
+            data['vendorid_set-TOTAL_FORMS'] = 3
+            data['vendorid_set-INITIAL_FORMS'] = 0
+            data['vendorid_set-MAX_NUM_FORMS'] = 0
+            response = self.client.post('%sinstdata/dataid/add/' % ADMIN_URL_PREFIX, data)
+            self.assertEqual(response.status_code,302)
+            id = self.get('DBR54')
+            self.assertEqual(id.type,'bond')